Fix: Add seq 카운터 복원 + 이벤트 정리 + ccusage 파싱 수정
- 서버 재시작 시 DB에서 세션별 최대 seq 복원 (이벤트 시퀀스 연속성)
- 24시간 이전 오래된 이벤트 자동 정리 (cleanup_old_events)
- ccusage blocks 응답 키 수정 (data -> blocks)
- burn_rate 타입 int -> float 수정
- 테스트 conftest.py에 공용 fixture 작성

+++ backend/app/schemas/usage.py
@@ -10,7 +10,7 @@ class BlockUsage(BaseModel):
     cost_usd: float = 0.0
     is_active: bool = False
     time_remaining: str = ""
-    burn_rate: int = 0
+    burn_rate: float = 0.0
 
 
 class WeeklyUsage(BaseModel):
